feat: 배송 - 배송과 배송 경로 생성 #33

+++ delivery/src/main/java/com/zgzg/delivery/application/dto/res/DeliveryRouteResponseDTO.java
@@ -1,6 +1,5 @@
 package com.zgzg.delivery.application.dto.res;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -20,10 +19,10 @@ public class DeliveryRouteResponseDTO {
 	private String startHubName;
 	private UUID endHubId;
 	private String endHubName;
-	private BigDecimal estimatedDistance;
-	private BigDecimal estimatedTime;
-	private BigDecimal actualDistance;
-	private BigDecimal actualTimeSpent;
+	private Integer estimatedDistance;
+	private Integer estimatedTime;
+	private Integer actualDistance;
+	private Integer actualTimeSpent;
 	private DeliveryStatus deliveryStatus;
 	private UUID deliveryPersonId;
 	private String deliveryPersonName;
